+++ test/data/observatory/test_dataset.py
@@ -12,27 +12,12 @@ from cartoframes.data.observatory.repository.variable_group_repo import Variable
 from cartoframes.data.observatory.repository.dataset_repo import DatasetRepository
 from .examples import test_dataset1, test_datasets, test_variables, test_variables_groups, db_dataset1, test_dataset2, \
     db_dataset2
+from .mocks import BigQueryClientMock, CredentialsMock
 
 try:
     from unittest.mock import Mock, patch
 except ImportError:
     from mock import Mock, patch
-
-
-class BigQueryClientMock(object):
-    def __init__(self, response):
-        self.response = response
-
-    def download_to_file(self, _1, _2, _3):
-        if isinstance(self.response, Exception):
-            raise self.response
-        else:
-            return self.response
-
-
-class CredentialsMock(object):
-    def __init__(self, username):
-        self.username = username
 
 
 class TestDataset(unittest.TestCase):
